Update AndroidManifest.xml
http://developer.android.com/guide/topics/data/install-location.html
The app can now be installed on the SDcard.
This was a feature-request.

Google Play app-review:
Jayvee Diaz on 11 oct. 2013 at 02:05
Nice and simple	I like how simple it looks just to keep the memory use down, but that's also it's biggest downside. It's so simple it's actually rather bland and boring. Not really a problem for what it does but it's rather underwhelming. And is there a reason why it can't be moved to an SD card?

The reason was:
The app could not be installed on the SDcard because
android:installLocation="preferExternal"
was not present in the AndroidManifest.xml